Create FlatteningAPI
Functional API with Flatten applied to output layers (of each primary branch model, i.e. outpute of Enhancer Model and outputp of Promoter Model). In this version, the Model is instantiated only once - this is in the concatenated model, as opposed to doing this in both Enhancer Model and Promoter Model.